Consumer/requirements.txt
kafka-python
mysql-connector-python
datetime